preview: extract citation-render + query-blocks from Preview.svelte (#1087)
The last piece of #1087 — pull the two concerns the issue names out of the
2968-line Preview.svelte into sibling modules under preview/ (matching the
existing text/image-paths/live-blocks/cite-meta convention):

- preview/citation-render.ts (171 LOC): the cite/quote DOM enrichment —
  applyCslMarkers (batch CSL-marker pass), resolveCiteLabel / resolveQuoteLabel
  (per-link graph metadata → tooltip). Takes a CitationRenderDeps { previewEl,
  cite/quoteMetaCache, queryPrefixes, setBibliographyEntries }; reads api /
  cite-meta directly.
- preview/query-blocks.ts (245 LOC): the query-block / chart rendering —
  executeQueryBlock (SPARQL/SQL + read-only backlinks/semantic live blocks) and
  the list/table/timeseries renderers. Takes QueryBlockDeps { notePath, revision,
  queryCache, queryPrefixes, activeCharts }; pushes chart handles into the
  caller's activeCharts (the component still destroys them before re-render).

Preview builds a fresh deps bundle per call (capturing current previewEl/props)
and passes it at each call site in the post-render $effect + transclusion
hydration. The injected HTML stays in the component's DOM, so the :global query/
cite CSS is untouched. Behavior identical.

Preview.svelte: 2968 → 2622. pnpm lint (0 errors) + full pnpm test (4043) green.
